Agregando menu variables aleatorias

+++ IntelliJ/Pseudoaleatorio/src/sample/Controllers/ventanaMixtoController.java
@@ -6,10 +6,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import sample.Main;
@@ -21,13 +18,11 @@ import java.net.URL;
 import java.util.ResourceBundle;
 
 public class ventanaMixtoController implements Initializable {
-    @FXML JFXButton btnRegresar;    @FXML JFXButton btnCerrar;
+    @FXML JFXButton btnRegresar;     @FXML JFXButton btnCerrar;
     @FXML JFXButton btnGenerar;      @FXML JFXButton btnPruebas;
     @FXML JFXTextField textCantidad;
-    @FXML
-    TableView<NumePseudoaleatorios> tbNumePseudoaleatorios;
-    @FXML
-    TableColumn clmNumePseudoaleatorios;
+    @FXML TableView<NumePseudoaleatorios> tbNumePseudoaleatorios;
+    @FXML TableColumn clmNumePseudoaleatorios;
 
     Main main=new Main();
     NumePseudoaleatorios modeloNumeros=null;
@@ -35,7 +30,7 @@ public class ventanaMixtoController implements Initializable {
     ObservableList<NumePseudoaleatorios> numeros=FXCollections.observableArrayList();
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-        clmNumePseudoaleatorios.setCellValueFactory(new PropertyValueFactory<NumePseudoaleatorios, Integer>("numeroPseudoaleatorio"));
+        clmNumePseudoaleatorios.setCellValueFactory(new PropertyValueFactory<NumePseudoaleatorios, Double>("numeroPseudoaleatorio"));
         tbNumePseudoaleatorios.setItems(llenarTabla());
         btnRegresar.setOnAction(event -> {
             Controller controller=new Controller();
